Add Form Requests & Improve Code

+++ app/Services/CommentService.php
@@ -4,6 +4,7 @@ namespace App\Services;
 
 use App\Models\Comment;
 use App\Models\Post;
+use Gate;
 
 /**
  * Class CommentService.
@@ -17,13 +18,20 @@ class CommentService
     }
     public static function update($commentId, $data)
     {
+        unset($data["user_id"]);
+        unset($data["post_id"]);
+
         $comment = Comment::find($commentId);
+        Gate::authorize("update", $comment);
+
         $comment->update($data);
         return $comment;
     }
     public static function destroy($commentId)
     {
         $comment = Comment::find($commentId);
+        Gate::authorize("delete", $comment);
+
         $comment->delete();
         return $comment;
     }
